more reorganization, now the PR is completely integrated
