+++ src/methods/methods.jl
@@ -173,10 +173,10 @@ end
 include("initial_guess.jl")
 include("differentials.jl")
 include("VT.jl")
+include("fugacity_coefficient.jl")
 include("property_solvers/property_solvers.jl")
 include("tpd.jl")
 include("stability.jl")
 include("pT.jl")
 include("unitful_base.jl")
 include("unitful_methods.jl")
-include("michelsen/michelsen.jl")
