+++ src/methods/michelsen/michelsen.jl
@@ -1,4 +1,1 @@
-import LinearAlgebra: I as Identity
-
 include("fugacity_coefficient.jl")
-include("saturation_pure.jl")
